fibonacci integer overflow working now

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,8 +5,8 @@
  */
 int main(void)
 {
-	unsigned long i;
-	int arr[50];
+	int i;
+	long long arr[50];
 
 	arr[0] = 1;
 	arr[1] = 2;
@@ -17,7 +17,7 @@ int main(void)
 	}
 	for (i = 0; i < 50; i++)
 	{
-		printf("%d", arr[i]);
+		printf("%lld", arr[i]);
 		if (i < 49)
 		{
 			printf(", ");
